+++ pakyow-core/lib/pakyow/behavior/running.rb
@@ -79,6 +79,8 @@ module Pakyow
         end
 
         def run
+          boot
+
           Async::Reactor.run do |reactor|
             @__reactor = reactor